backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
email-validator==2.2.0
docling
pymupdf
numpy
google-generativeai
google-genai
chromadb
sentence-transformers
einops
huggingface_hub
llama-cpp-python
